app/requirements.txt
flask
mysql-connector
pyopenssl
gunicorn
